Modified offset member of structure dpiErrorInfo to be 32-bit instead of 16-bit in order to allow for row offsets that exceed 65536. The original member has been renamed to offset16 in order to retain backwards compatibility (https://github.com/oracle/node-oracledb/issues/1157).

+++ src/dpiImpl.h
@@ -780,7 +780,7 @@ typedef struct {
 // with the function dpiStmt_getBatchErrors()
 typedef struct {
     int32_t code;                       // Oracle error code or 0
-    uint16_t offset;                    // parse error offset or row offset
+    uint32_t offset;                    // parse error offset or row offset
     dpiErrorNum errorNum;               // OCPI-C error number
     const char *fnName;                 // ODPI-C function name
     const char *action;                 // internal action
